+++ app/helpers/early_career_payments_helper.rb
@@ -4,8 +4,6 @@ module EarlyCareerPaymentsHelper
       content_tag(:h1, I18n.t("early_career_payments.ineligible.poor_performance_heading"), class: "govuk-heading-xl")
     elsif claim.eligibility.ineligibility_reason == :ineligible_current_school
       content_tag(:h1, I18n.t("early_career_payments.ineligible.school_heading"), class: "govuk-heading-xl")
-    elsif claim.eligibility.ineligibility_reason == :ineligible_nqt_in_academic_year_after_itt
-      content_tag(:h1, I18n.t("early_career_payments.ineligible.reason.nqt_after_itt"), class: "govuk-heading-l")
     else
       content_tag(:h1, I18n.t("early_career_payments.ineligible.heading"), class: "govuk-heading-xl")
     end
@@ -13,5 +11,13 @@ module EarlyCareerPaymentsHelper
 
   def one_time_password_validity_duration
     pluralize(OneTimePassword::Base::DRIFT / 60, "minute")
+  end
+
+  def eligible_itt_subject_translation(claim)
+    if claim.eligibility.trainee_teacher_in_2021?
+      I18n.t("early_career_payments.questions.eligible_itt_subject_trainee_teacher_in_2021")
+    else
+      I18n.t("early_career_payments.questions.eligible_itt_subject", qualification: claim.eligibility.qualification_name)
+    end
   end
 end