+++ app/models/claim_stats.rb
@@ -1,0 +1,11 @@
+class ClaimStats < ApplicationRecord
+  self.primary_key = "claim_id"
+
+  def readonly?
+    true
+  end
+
+  def self.refresh
+    Scenic.database.refresh_materialized_view(table_name, concurrently: false, cascade: false)
+  end
+end