+++ lib/student_loan.rb
@@ -14,12 +14,19 @@ module StudentLoan
     BEFORE_AND_AFTER_1_SEPT_2012 = "some_before_some_after_first_september_2012"
   ].freeze
 
-  PLAN_1_COUNTRIES = [NORTHERN_IRELAND, SCOTLAND].freeze
+  PLAN_1_COUNTRIES = [NORTHERN_IRELAND].freeze
+  PLAN_4_COUNTRIES = [SCOTLAND].freeze
 
   PLANS = [
     PLAN_1 = "plan_1",
     PLAN_2 = "plan_2",
-    PLAN_1_AND_2 = "plan_1_and_2"
+    PLAN_1_AND_2 = "plan_1_and_2",
+    PLAN_4 = "plan_4",
+    PLAN_3 = "plan_3",
+    PLAN_1_AND_3 = "plan_1_and_3",
+    PLAN_2_AND_3 = "plan_2_and_3",
+    PLAN_1_AND_2_AND_3 = "plan_1_and_2_and_3",
+    PLAN_4_AND_3 = "plan_4_and_3"
   ].freeze
 
   DATES_TO_PLANS_MAP = {
@@ -29,13 +36,27 @@ module StudentLoan
   }.freeze
 
   # Used to determine a person's student loan plan based on their country of
-  # study and the start date(s) of their course(s).
+  # study and the start date(s) of their course(s) if they have a student loan.
   #
   # Returns nil if the plan cannot be determined based on the information
   # provided.
-  def self.determine_plan(country, course_start_date = nil)
-    return PLAN_1 if PLAN_1_COUNTRIES.include?(country)
+  # Or returns PLAN_3 when has a postgraduate masters and/or doctoral loan
+  def self.determine_plan(has_student_loan, has_postgraduate_loan, country = nil, course_start_date = nil)
+    return Claim::NO_STUDENT_LOAN if !has_student_loan && !has_postgraduate_loan
+    return PLAN_3 if !has_student_loan && has_postgraduate_loan
 
-    DATES_TO_PLANS_MAP[course_start_date]
+    base_plan = case country
+    when SCOTLAND
+      PLAN_4
+    when NORTHERN_IRELAND
+      PLAN_1
+    else
+      DATES_TO_PLANS_MAP[course_start_date]
+    end
+
+    return base_plan unless has_postgraduate_loan
+    return base_plan if base_plan.nil?
+
+    const_get([base_plan.upcase, :_AND_3].join)
   end
 end